+++ tests/ConnectionManagerTest.php
@@ -46,6 +46,45 @@ class ConnectionManagerTest extends TestCase
         );
     }
 
+    public function testConnectionsTwoSockets()
+    {
+        $serverManager = new ConnectionManager(
+            [
+                'tcp://127.0.0.1:11945',
+                'tcp://127.0.0.1:11946',
+            ],
+            new NullLogger(),
+            new TestSocket()
+        );
+
+        $this->assertSame(
+            [
+                [
+                    'common_name' => 'f3bb6f8efb4dc64be35e1044cf1b5e76',
+                    'virtual_address' => [
+                        '10.128.7.3',
+                        'fd60:4a08:2f59:ba0::1001',
+                    ],
+                ],
+                [
+                    'common_name' => '78f4a3c26062a434b01892e2b23126d1',
+                    'virtual_address' => [
+                        '10.128.7.4',
+                        'fd60:4a08:2f59:ba0::1002',
+                    ],
+                ],
+                [
+                    'common_name' => '67a7e629c4112b4a85fb254660129f2c',
+                    'virtual_address' => [
+                        '10.104.9.130',
+                        'fdd3:1503:4c0e:1da1::1000',
+                    ],
+                ],
+            ],
+            $serverManager->connections()
+        );
+    }
+
     public function testConnectionsNoConnections()
     {
         $serverManager = new ConnectionManager(