+++ CHANGES.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## 1.0.2 (2018-06-13)
+- fix bug where connections where not properly aggregated when using multiple
+  socket addresses
+
 ## 1.0.1 (2018-06-07)
 - add description in `composer.json`